requirements.txt
weaviate-client
fastapi
uvicorn
openai
qdrant-client
litellm
pydantic
motor